+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+### Changes
+
+- React Native Tracing Deprecations  ([#4073](https://github.com/getsentry/sentry-react-native/pull/4073))
+  - `new ReactNativeTracing` to `reactNativeTracingIntegration()`
+  - `new ReactNavigationInstrumentation` to `reactNativeTracingIntegration()`.
+  - `new ReactNativeNavigationInstrumentation` to `reactNativeTracingIntegration()`.
+  - `ReactNavigationV4Instrumentation` won't be supported in the next major SDK version, upgrade to `react-navigation@5` or newer.
+  - `RoutingInstrumentation` and `RoutingInstrumentationInstance` replace by `Integration` interface from `@sentry/types`.
+  - `enableAppStartTracking`, `enableNativeFramesTracking`, `enableStallTracking`, `enableUserInteractionTracing` moved to `Sentry.init({})` root options.
+
 ## 5.31.1
 
 ### Fixes
@@ -632,7 +644,7 @@ This release contains upgrade of `sentry-android` dependency to major version 7.
 ## 5.16.0
 
 This release ships with a beta version of our new built-in Expo SDK 50 support,
-which replaces the deprecated `sentry-expo` package. To learn more,
+which replaces the `sentry-expo`package. To learn more,
 see [the Expo guide](https://docs.sentry.io/platforms/react-native/manual-setup/expo/).
 
 ### Features
@@ -1012,7 +1024,7 @@ This release is compatible with `expo@50.0.0-preview.6` and newer.
 - `Sentry.wrap` doesn't enforce any keys on the wrapped component props ([#3332](https://github.com/getsentry/sentry-react-native/pull/3332))
 - Ignore defaults when warning about duplicate definition of trace propagation targets ([#3327](https://github.com/getsentry/sentry-react-native/pull/3327))
 - Screenshots are not taken when the SDK is disabled ([#3333](https://github.com/getsentry/sentry-react-native/pull/3333))
-- Use deprecated `ReactNativeTracingOptions.tracingOrigins` if set in the options ([#3331](https://github.com/getsentry/sentry-react-native/pull/3331))
+- Use `ReactNativeTracingOptions.tracingOrigins`if set in the options ([#3331](https://github.com/getsentry/sentry-react-native/pull/3331))
 - Cancel auto instrumentation transaction when app goes to background ([#3307](https://github.com/getsentry/sentry-react-native/pull/3307))
 
 ### Dependencies